creates string validator with working specs
