+++ lib/app.rb
@@ -7,3 +7,4 @@ require 'rgb'
 require 'rgba'
 require 'rgb_a_utils'
 require 'rgb_value_parser'
+require 'string_validator'
